Reverting changes from revisions #9874 and #9873.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9877 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/dparser/arg.c
@@ -59,7 +59,7 @@ process_arg(ArgumentState *arg_state, int i, char ***argv) {
 
 void
 process_args(ArgumentState *arg_state, char **argv) {
-  uint i = 0, len;
+  int i = 0, len;
   char *end;
   ArgumentDescription *desc = arg_state->desc;
   /* Grab Environment Variables */
@@ -136,7 +136,7 @@ process_args(ArgumentState *arg_state, char **argv) {
 void 
 usage(ArgumentState *arg_state, char *arg_unused) {
   ArgumentDescription *desc = arg_state->desc;
-  uint i;
+  int i;
 
   (void)arg_unused;
   fprintf(stderr,"Usage: %s [flags|args]\n",arg_state->program_name);
@@ -150,8 +150,8 @@ usage(ArgumentState *arg_state, char *arg_unused) {
 	    desc[i].key != ' ' ? ',' : ' ', desc[i].name,
             (strlen(desc[i].name) + 61 < 81) ?
              &SPACES[strlen(desc[i].name)+61] : "",
-            arg_types_desc[desc[i].type?(uint)(strchr(arg_types_keys,desc[i].type[0])-
-                arg_types_keys) : strlen(arg_types_keys)]);
+            arg_types_desc[desc[i].type?strchr(arg_types_keys,desc[i].type[0])-
+                arg_types_keys : strlen(arg_types_keys)]);
     switch(desc[i].type?desc[i].type[0]:0) {
       case 0: fprintf(stderr, "          "); break;
       case 'L':
